refactor : Test Configuration 공통부분을 별도 클래스로 추출한다
