+++ src/test/java/org/prgrms/kdt/TestConfiguration.java
@@ -1,0 +1,32 @@
+package org.prgrms.kdt;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+
+import javax.sql.DataSource;
+
+@Configuration
+@ComponentScan(
+        basePackages = {"org.prgrms.kdt.repository", "org.prgrms.kdt.service"}
+)
+public class TestConfiguration {
+
+    @Bean
+    public DataSource dataSource() {
+        return DataSourceBuilder.create()
+                .url("jdbc:mysql://localhost:2215/test-order-mgmt")
+                .username("test")
+                .password("test1234!")
+                .type(HikariDataSource.class)
+                .build();
+    }
+
+    @Bean
+    public NamedParameterJdbcTemplate jdbcTemplate(DataSource dataSource) {
+        return new NamedParameterJdbcTemplate(dataSource);
+    }
+}
